logosimilarity/requirements.txt
pyspark~=4.0.1
ImageHash~=4.3.2
pillow~=12.0.0
requests~=2.32.5